refactor: 迁移 moduleManager/ → modules/ (Phase 4a)
目录重组第一步：将模块管理器从 src/moduleManager/ 迁移到 src/modules/。
- 21 个源文件通过 git mv 重命名（保留 git 历史）
- extension.ts 导入路径更新
- 11 个测试文件导入路径更新
- src/moduleManager/index.ts 保留 re-export 兼容层

内部导入（./xxx）不变，仅更新外部引用。352 测试全部通过。
