+++ src/modules/labviewVersionDetector.ts
