+++ src/modules/moduleSidebarHtml.ts
