+++ src/modules/moduleSidebarViewProvider.ts
